Refactored the play logic into a separate class.
