+++ BrickController2/BrickController2/CreationManagement/ControllerAxisType.cs
@@ -3,6 +3,7 @@
     public enum ControllerAxisType
     {
         Normal,
-        Train
+        Train,
+        Accelerator
     }
 }
