+++ BrickController2/BrickController2/CreationManagement/ControllerButtonType.cs
@@ -7,6 +7,7 @@
         Alternating,
         Circular,
         PingPong,
-        Stop
+        Stop,
+        Accelerator
     }
 }
